requirements.txt
# Core dependencies
python-dotenv>=0.19.0
gcsa>=1.3.0
beautiful-date>=2.0.0
pytz>=2023.3

# You may also need to install the linak-controller utility separately
# This is typically installed via pip or your system package manager